s/uint64_t/double/ to avoid implicit cast return.

+++ tiledb/sm/query/readers/sparse_index_reader_base.h
@@ -160,7 +160,7 @@ class MemoryBudget {
    * @return Portion of the total memory budget dedicated to loading coordinate
    * tiles.
    */
-  uint64_t coordinates_budget() {
+  double coordinates_budget() {
     return total_budget_ * ratio_coords_;
   }
